remove unused function and move UT
