+++ presidio-analyzer/tests/test_entity_recognizer.py
@@ -19,20 +19,6 @@ def test_when_from_dict_then_returns_instance():
     assert entity_rec.version == "0.0.1"
 
 
-def test_when_index_finding_then_succeed():
-    # This test uses a simulated recognize result for the following
-    # text: "my phone number is:(425) 882-9090"
-    match = "(425) 882-9090"
-    # the start index of the match
-    start = 19
-    tokens = ["my", "phone", "number", "is:(425", ")", "882", "-", "9090"]
-    tokens_indices = [0, 3, 9, 16, 23, 25, 28, 29]
-    index = EntityRecognizer._find_index_of_match_token(
-        match, start, tokens, tokens_indices
-    )
-    assert index == 3
-
-
 def test_when_remove_duplicates_duplicates_removed():
     # test same result with different score will return only the highest
     arr = [
