+++ presidio-analyzer/tests/test_lemma_context_aware_enhancer.py
@@ -1,0 +1,15 @@
+from presidio_analyzer import LemmaContextAwareEnhancer
+
+
+def test_when_index_finding_then_succeed():
+    # This test uses a simulated recognize result for the following
+    # text: "my phone number is:(425) 882-9090"
+    match = "(425) 882-9090"
+    # the start index of the match
+    start = 19
+    tokens = ["my", "phone", "number", "is:(425", ")", "882", "-", "9090"]
+    tokens_indices = [0, 3, 9, 16, 23, 25, 28, 29]
+    index = LemmaContextAwareEnhancer._find_index_of_match_token(
+        match, start, tokens, tokens_indices
+    )
+    assert index == 3
